write method update
