+++ demisto_sdk/commands/common/files/json_file.py
@@ -1,12 +1,10 @@
-from functools import lru_cache
-from io import BytesIO
-from typing import Optional, Set, Union
+from typing import Set
 
 from pydantic import Field, validator
 
 from demisto_sdk.commands.common.files.handler_file import HandlerFile
 from demisto_sdk.commands.common.handlers import DEFAULT_JSON_HANDLER as json
-from demisto_sdk.commands.common.handlers import JSON_Handler, XSOAR_Handler
+from demisto_sdk.commands.common.handlers import JSON_Handler
 
 
 class JsonFile(HandlerFile):
@@ -19,14 +17,3 @@ class JsonFile(HandlerFile):
     @classmethod
     def known_extensions(cls) -> Set[str]:
         return {".json"}
-
-    @classmethod
-    @lru_cache
-    def read_from_file_content(
-        cls,
-        file_content: Union[bytes, BytesIO],
-        handler: Optional[XSOAR_Handler] = None,
-    ):
-        return super().read_from_file_content(
-            file_content, handler=cls.validate_handler()
-        )
